Some refactoring and new projects

+++ benchmark/prepare_project.py
@@ -10,7 +10,7 @@ from coder.utils import run_query
 
 CURSOR = '<CURSOR>'
 
-def prepare(config, mode, ids=[]):
+def prepare(config, mode, ids=[], noTests=False):
     global CURSOR
 
     here = Path(__file__).resolve().parent
@@ -25,8 +25,9 @@ def prepare(config, mode, ids=[]):
         dir_util.mkpath(str(here/'experiment'/config['name']/mode/f'temp{i["id"]}'))
         temp_dir = here/'experiment'/config['name']/mode/f'temp{i["id"]}'/config['project_root']
         dir_util.copy_tree(str(here/config['project_root']), str(temp_dir))
-        dir_util.remove_tree(str(temp_dir/'tests'))
-        if i['id'] == 0:
+        dir_util.remove_tree(str(temp_dir/config['tests_path']))
+        orig_results = {"tests": 0, "errors": 0, "failures": 0, "skipped": 0, "id": 0}
+        if i['id'] == 0 and not noTests:
             orig_results = run_tests(config, 0, mode, env_session.interpreter.executable)
         if len(i['file']) == 0:
             continue
@@ -39,44 +40,46 @@ def prepare(config, mode, ids=[]):
         with open(temp_dir/i["file"]) as f:
             code = f.readlines()
         new_code = []
-        cursor = 'with open("' + str(here/'experiment'/config['name']/mode/f'temp{i["id"]}'/'checkTest.txt') + '", "w") as f: f.write("here")\n'
-        for l in range(len(code)):
-            temp = code[l]
-            for j in i['remove']:
-                if j['description'] == 'imports':
-                    continue
-                if j['start_line'] - 1 <= l <= j['end_line'] - 1:
-                    if j['start_line'] == j['end_line']:
-                        temp = code[l][:j['start_column']] + (cursor if j['description'] != 'imports' else '') + code[l][j['end_column'] - 1:]
-                    else:
-                        if l == j['start_line'] - 1:
-                            temp = code[l][:j['start_column']] + (cursor if j['description'] != 'imports' else '')
-                        elif l == j['end_line'] - 1:
-                            temp = code[l][j['end_column'] - 1:]
+        if not noTests:
+            cursor = 'with open("' + str(here/'experiment'/config['name']/mode/f'temp{i["id"]}'/'checkTest.txt') + '", "w") as f: f.write("here")\n'
+            for l in range(len(code)):
+                temp = code[l]
+                for j in i['remove']:
+                    if j['description'] == 'imports':
+                        continue
+                    if j['start_line'] - 1 <= l <= j['end_line'] - 1:
+                        if j['start_line'] == j['end_line']:
+                            temp = code[l][:j['start_column']] + (cursor if j['description'] != 'imports' else '') + code[l][j['end_column'] - 1:]
                         else:
-                            temp = None
-            if temp:
-                new_code.append(temp)
-        with open(temp_dir/i["file"], 'w') as f:
-            f.writelines(new_code)
-        test_results = run_tests(config, i['id'], mode, env_session.interpreter.executable)
-        try:
-            with open(str(here/'experiment'/config['name']/mode/f'temp{i["id"]}'/'checkTest.txt'), 'r') as f:
-                content = f.read()
-        except:
-            dir_util.remove_tree(str(here/'experiment'/config['name']/mode/f'temp{i["id"]}'))
-            continue
-        if content != 'here':
-            dir_util.remove_tree(str(here/'experiment'/config['name']/mode/f'temp{i["id"]}'))
-            continue
-        different = False
-        for k, v in test_results.items():
-            if k not in orig_results or orig_results[k] != v:
-                different = True
-                break
-        if not different:
-            dir_util.remove_tree(str(here/'experiment'/config['name']/mode/f'temp{i["id"]}'))
-            continue
+                            if l == j['start_line'] - 1:
+                                temp = code[l][:j['start_column']] + (cursor if j['description'] != 'imports' else '')
+                            elif l == j['end_line'] - 1:
+                                temp = code[l][j['end_column'] - 1:]
+                            else:
+                                temp = None
+                if temp:
+                    new_code.append(temp)
+            with open(temp_dir/i["file"], 'w') as f:
+                f.writelines(new_code)
+            
+            test_results = run_tests(config, i['id'], mode, env_session.interpreter.executable)
+            try:
+                with open(str(here/'experiment'/config['name']/mode/f'temp{i["id"]}'/'checkTest.txt'), 'r') as f:
+                    content = f.read()
+            except:
+                dir_util.remove_tree(str(here/'experiment'/config['name']/mode/f'temp{i["id"]}'))
+                continue
+            if content != 'here':
+                dir_util.remove_tree(str(here/'experiment'/config['name']/mode/f'temp{i["id"]}'))
+                continue
+            different = False
+            for k, v in test_results.items():
+                if k not in orig_results or orig_results[k] != v:
+                    different = True
+                    break
+            if not different:
+                dir_util.remove_tree(str(here/'experiment'/config['name']/mode/f'temp{i["id"]}'))
+                continue
         database = temp_dir/'..'/'..'/'codeqldb'
         working_dir = os.getcwd()
         os.chdir(str(temp_dir))
